+++ modules/spring-core/src/test/kotlin/com/esprito/spring/core/providers/java/SpringBeanLineMarkerProviderInheritanceTest.kt
@@ -34,7 +34,8 @@ class SpringBeanLineMarkerProviderInheritanceTest : EspritoJavaLightTestCase() {
         myFixture.configureFromExistingVirtualFile(vf)
         myFixture.doHighlighting()
 
-        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, SpringIcons.SpringBean)
+        val icons = setOf(SpringIcons.SpringBean, SpringIcons.springBeanInactive)
+        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, icons)
         val gutterTargetString = getGutterTargetString(allBeanGutters)
 
         TestCase.assertEquals(gutterTargetString.flatMap { gutter ->
@@ -77,7 +78,8 @@ class SpringBeanLineMarkerProviderInheritanceTest : EspritoJavaLightTestCase() {
         myFixture.configureFromExistingVirtualFile(vf)
         myFixture.doHighlighting()
 
-        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, SpringIcons.SpringBean)
+        val icons = setOf(SpringIcons.SpringBean, SpringIcons.springBeanInactive)
+        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, icons)
         val gutterTargetString = getGutterTargetString(allBeanGutters)
 
         TestCase.assertEquals(gutterTargetString.flatMap { gutter ->
@@ -120,7 +122,8 @@ class SpringBeanLineMarkerProviderInheritanceTest : EspritoJavaLightTestCase() {
         myFixture.configureFromExistingVirtualFile(vf)
         myFixture.doHighlighting()
 
-        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, SpringIcons.SpringBean)
+        val icons = setOf(SpringIcons.SpringBean, SpringIcons.springBeanInactive)
+        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, icons)
         val gutterTargetString = getGutterTargetString(allBeanGutters)
 
         TestCase.assertEquals(gutterTargetString.flatMap { gutter ->
@@ -163,7 +166,8 @@ class SpringBeanLineMarkerProviderInheritanceTest : EspritoJavaLightTestCase() {
         myFixture.configureFromExistingVirtualFile(vf)
         myFixture.doHighlighting()
 
-        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, SpringIcons.SpringBean)
+        val icons = setOf(SpringIcons.SpringBean, SpringIcons.springBeanInactive)
+        val allBeanGutters = getAllBeanGuttersByIcon(myFixture, icons)
         val gutterTargetString = getGutterTargetString(allBeanGutters)
 
         TestCase.assertEquals(gutterTargetString.flatMap { gutter ->